refactor: remove dev monkey parch from source
